Improved build script. Additionally now can load custom HTTPs certificates.

+++ Backend/build.sh
@@ -1,13 +1,31 @@
 #!/bin/bash
 cd "$(dirname "$0")"
 
-dotnet publish
+RED='\033[0;31m'
+GREEN='\033[0;32m'
+NC='\033[0m' # No Color
 
-#dotnet publish --os linux --arch x64
+target_os=$1
+target_arch=$2
+
+if [[ "$target_os" == "" ]] || [[ "$target_arch" == "" ]]
+then
+  echo "Both Target OS as well as arch must be specified!"
+  printf "\nExample usage: ./build.sh linux x64"
+  exit -1
+fi
+
+dotnet_result=$(dotnet publish --configuration release --os $target_os --arch $target_arch)
+
+if [[ "$dotnet_result" == *"error"* ]];
+then
+  printf "${RED}\nBuild Error Detected: \n ${dotnet_result} ${NC}"
+  exit -1
+fi
 
 rm -rf bin/out # clean previous output
 
-cp bin/Debug/net6.0/publish bin/out -R # make our own folder for output, leave Debug be
+cp bin/Release/net6.0/${target_os}-${target_arch}/publish bin/out -R # make our own folder for output
 
 # clear any previous state
 rm bin/out/monopolystate.json
@@ -26,22 +44,24 @@ STATIC_PATH=./static
 DEVELOPMENT_CORS=https://192.168.0.69:3000
 " > bin/out/.env
 
-## copy NLog production instead of dev one
-# cp NLog.production.config bin/out/NLog.config
-
 ## Generate frontend files and move them to folder
 cd ../Frontend/
 if [ -d "build" ]; then
-    echo "build React output folder already exists! Simply copying. Delete and rerun to rebuild it or rebuild it manually."
+    printf "\n Build React output folder already exists! Simply copying. Delete and rerun to rebuild it or rebuild it manually. \n"
 else
-    echo "Build folder from react is not present, building..."
+    printf "\n Build folder from react is not present, building... \n "
     npm run build
 fi
 
 cp build ../Backend/bin/out/static -R
 fullpath=$(realpath ../Backend/bin/out)
+fullpath_settings=$(realpath ../Backend/bin/out/appsettings.json)
+keypath=$(realpath ../Backend/bin/out/key.pem)
+certpath=$(realpath ../Backend/bin/out/cert.pem)
 
+printf "\nConfiguration for Server settings can be set in ${fullpath_settings}\n"
+printf "\nYou will need a certificate to run with current appsettings.json config"
+printf "\nEither delete the certificate configuration and switch to HTTP"
+printf "\nOr generate a key.pem and cert.pem using the command: \n\n openssl req -x509 -newkey rsa:4096 -keyout bin/out/key.pem -out bin/out/cert.pem -sha256 -days 365 -nodes \n"
 
-echo "\n This build was compiled for linux x64 systems. Modify the first line of build.sh to change this\n"
-echo "Build Finished. See --> ${fullpath}"
-
+printf "${GREEN}\nBuild Finished. See --> ${fullpath} ${NC}"
